+++ NEWS.md
@@ -12,4 +12,4 @@
 
 # dca 0.1.0
 
-* First release of the R package to implement Decision Curve Analysis. (decisioncurveanalysis.org)
+* First release of the R package to implement Decision Curve Analysis.