+++ NEWS.md
@@ -1,3 +1,5 @@
+# dcurves (development version)
+
 # dcurves 0.2.0
 
 * Changed name of package from dca to dcurves.